Use concerns to share instance methods
